Rename MainContent
