+++ src/components/App/__tests__/Main.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { create } from 'react-test-renderer';
 
-import MainContent from '../MainContent';
+import Main from '../MainContent';
 
-describe('COMPONENT - App MainContent', () => {
+describe('COMPONENT - App Main', () => {
   it('renders correctly', () => {
-    const component = create(<MainContent />);
+    const component = create(<Main />);
 
     expect(component.toJSON()).toMatchSnapshot();
   });
